requirements.txt
streamlit>=1.32.0
plotly>=5.20.0
pandas>=2.2.0
httpx>=0.27.0
python-binance>=1.0.35
python-dotenv>=1.0.0
